+++ src/evo/deterministicmns.cpp
@@ -13,7 +13,6 @@
 #include <base58.h>
 #include <chainparams.h>
 #include <consensus/validation.h>
-#include <core_io.h>
 #include <script/standard.h>
 #include <ui_interface.h>
 #include <validation.h>
